Move the mutation API out to separate builder objects.
Here's a rough proposal for how it could look to move the mutation side
of the macro API to objects that are passed separately to the macro.
This means there is no longer a need to maintain separate pairs of Foo
and TargetFoo classes.

I don't think it's a huge win in this PR, but as these APIs fill out
with more members, I think the separation would become more valuable.